// types/ParagraphBlockReferenceProps.ts
export type ParagraphBlockReferenceProps = {
    __typename: 'ParagraphBlockReference'
    id: string
    bgColor: string
    bgImage: {
        mediaImage2: {
            url: string
        }
    }
    bgImagePosition: string
    bgImageRepeat: string
    bgImageSize: string
    fullWidth: boolean
    margin: string
    padding: string
    sectionId: string
    blockReference: {
        id: string
        render: string
    }
}